Move lib namespace to Nintron27/pillow
Move the NATS-pillow lib namespace from Nintron27/nats-pillow to
Nintron27/pillow
