+++ README.md
@@ -1,10 +1,20 @@
-![Nats Pillow icon](https://github.com/user-attachments/assets/7fd40216-2a5d-4ca8-b53e-04710038e718)
+![Pillow icon](https://github.com/user-attachments/assets/7ed49aab-a998-4bef-80b4-230b6ad87690)
 
-# NATS Pillow
-A soft little wrapper for NATS, to remove the boilerplate for certain use cases, mainly running NATS embedded in Go.
+# Pillow
+A soft little wrapper for [NATS](https://nats.io/) ^-^
+Remove the boilerplate in your embedded NATS projects, and deploy to cloud platforms with added functionality!
 
 > [!WARNING]
-> The Pillow API is unstable until v1.0.0 and as such may be subject to change in future releases until v1.0.0 is released.
+> The Pillow API is unstable until v1.0.0 and as such may be subject to change in future releases until v1.0.0 is reached.
+
+## Getting Started
+For simply embedding in a Go program:
+
+```shell
+go get github.com/Nintron27/pillow
+```
+
+then follow the [embedded example](./examples/embedded/embedded.go).
 
 ## Features
 - Remove boilerplate when using NATS embedded in Go
@@ -23,6 +33,7 @@ A soft little wrapper for NATS, to remove the boilerplate for certain use cases,
 - Flyio (Clustering & HubAndSpoke): Scaling down a JetStream enabled region may cause the JS quorum to fail, or the meta leader to not be established. This renders JS unuseable, and will hopefully be fixed or addressed in a future release.
 
 > [!CAUTION]
-> Take great caution when switching adapters (such as HubAndSpoke to Clustering) as the cluster names and JS domains will change. Along with scaling down JS regions, as there seems to be a bug with quorums failing.
+> Take great caution when switching adapters (such as HubAndSpoke to Clustering) as the cluster names and JS domains will change.
+> Also caution with scaling down JS regions, as there seems to be a bug with quorums failing.
 
 *Credit to [@whaaaley](https://github.com/whaaaley) for project name and icon*
