+++ adapters.go
